requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.74
botocore==1.34.74
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
Django==5.0.6
django-allow-cidr==0.7.1
django-celery-results==2.5.1
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-formtools==2.5.1
django-js-reverse==0.10.2
django-modeltranslation==0.18.11
django-ses==3.5.2
django-storages==1.14.2
django-webpack-loader==3.1.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
idna==3.7
jmespath==1.0.1
kombu==5.3.7
openpyxl==3.1.3
packaging==24.0
phonenumbers==8.13.38
pika==1.3.2
pillow==10.3.0
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.6
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
